Extract feature copy helpers to src/qwen35/feature_copy.{h,cpp}
Move target_capture_index, copy_capture_slice_to_draft_ring, and
copy_feature_ring_range_to_tensor from test_dflash.cpp into reusable
library files. copy_capture_slice_to_remote_draft stays in test (depends
on DFlashDraftIpcClient).
